Support cloning a SourceText with a content type
