+++ src/EditorFeatures/Core/ContentTypeNames.cs
@@ -1,16 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-namespace Microsoft.CodeAnalysis.Editor
-{
-    internal static class ContentTypeNames
-    {
-        public const string CSharpContentType = "CSharp";
-        public const string CSharpSignatureHelpContentType = "CSharp Signature Help";
-        public const string RoslynContentType = "Roslyn Languages";
-        public const string VisualBasicContentType = "Basic";
-        public const string VisualBasicSignatureHelpContentType = "Basic Signature Help";
-        public const string XamlContentType = "XAML";
-        public const string JavaScriptContentTypeName = "JavaScript";
-        public const string TypeScriptContentTypeName = "TypeScript";
-    }
-}
+using System.Runtime.CompilerServices;
+
+// Microsoft.CodeAnalysis.Editor.ContentTypeNames has been moved to Microsoft.CodeAnalysis.Editor.Text.dll
+[assembly: TypeForwardedTo(typeof(Microsoft.CodeAnalysis.Editor.ContentTypeNames))]
