+++ src/EditorFeatures/Test/Extensions/SourceTextContainerExtensionsTests.cs
@@ -28,9 +28,9 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Extensions
             bufferMock.SetupGet(x => x.CurrentSnapshot).Returns(textSnapshotMock.Object);
             bufferMock.SetupGet(x => x.Properties).Returns(new VisualStudio.Utilities.PropertyCollection());
 
-            var textContainer = Text.Extensions.TextBufferContainer.From(bufferMock.Object);
+            var textContainer = CodeAnalysis.Text.Extensions.TextBufferContainer.From(bufferMock.Object);
 
-            Text.Extensions.GetTextBuffer(textContainer);
+            CodeAnalysis.Text.Extensions.GetTextBuffer(textContainer);
         }
     }
 }
