+++ src/EditorFeatures/Text/Implementation/TextBufferFactoryService/ITextBufferCloneService.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using Microsoft.CodeAnalysis.Editor;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Utilities;
@@ -21,5 +22,15 @@ namespace Microsoft.CodeAnalysis.Text
         /// it is explicitly marked with unknown content type so that it can't be used with editor directly
         /// </summary>
         ITextBuffer CloneWithUnknownContentType(ITextImage textImage);
+
+        /// <summary>
+        /// get new <see cref="ITextBuffer"/> from <see cref="SourceText"/> with <see cref="ContentTypeNames.RoslynContentType"/>
+        /// </summary>
+        ITextBuffer CloneWithRoslynContentType(SourceText sourceText);
+
+        /// <summary>
+        /// get new <see cref="ITextBuffer"/> from <see cref="SourceText"/> with <see cref="IContentType"/>
+        /// </summary>
+        ITextBuffer Clone(SourceText sourceText, IContentType contentType);
     }
 }
